The HTML5 logo shield
This design is made of two paths, one for the darker main shield shape, and one for the lighter highlight on
the right half of the shield shape.

Let's look at the path definition for the darker part: [ M 0 0 H 182 L 165 185 L 90 206 L 17 185 Z ]

There are six instructions here:
 - M 0 0
 - H 182
 - L 165 185
 - L 90 206
 - L 17 185
 - Z

Think of these commands as moving an imaginary pen around the screen to draw lines.
 - The - M - command takes a position as an (x, y) coordinate pair and moves the pen to that position without drawing anything.
 - The - H - command takes an x-coordinate and draws a horizontal line from the current pen position to that value of x.
 - The - L - command takes an (x, y) coordinate pair and draws a line from the current position to that position.
 - Finally, the - Z - command closes the path, drawing a line from the current position back to that start of the path.

In English, the commands in the path say:
 - Move to (0, 0)
 - draw a horizontal line to (182, 0)
 - draw a line to (165, 185)
 - draw a line to (90, 206)
 - draw a line to (17, 185)
 - then draw a line back to (0, 0) to close the path

The second path uses the same technique to draw the inner highlight on the shield, using a different fill color.

These commands define the points to move to using - absolute positions - exact x- and y-coordinates.

There's an alternative version of each command that take a relative position instead, meaning that next point is defined in relation
to the current position of the pen.

The absolute commands all use uppercase letters, and the relative ones use the same letters but lowercase.

For example, the path definition we just looked at could be rewritten using relative path commands like this:
 [ m 0 0 h 182 l -17 185 l -75 21 l -73 -21 z ]

In this case, the - move - command is the same because there's no previous position to be relative to.
 - The command [ h 182 ] says to draw a horizontal line 182 units to the right of the current position.
 - The command [ l -17 185 ] says to draw a line 17 units to the left and 185 down from the current position, and so on.
 - The - Z - and - z - commands do the same thing and are just included in the SVG spec for completeness.

In fact, this relative path definition can be written even more compactly:
 [ m0 0h182l-17 185-75 21-73-21z ]
Spaces are needed only to avoid ambiguity between two numbers, but are otherwise optional in SVG paths.
 - Thanks to all the negative numbers, we're able to remove almost all the spaces here.

Also, if the same command is used multiple times in a row, you can inlcude it once and then just keep providing numbers.
 - For Example:
   - [ l-17 185-75 21-73-21 ] is the compact version of [ l -17 185 l -75 21 l -73 -21 ]