adding interactivity with JavaScript
We're using the - querySelectorAll - method to select all the elements with the class name [.fun] We then use
the - forEach - method to iterate over the selected elements and log them to the console.

When you run this code, you should see the three elements in the console on seperate lines.

If you hover over each elemnt in the console, that element should also be highlighted on the web page.