updated the - forEach - method call
We're doing two things to each element. First, we set something called a 'data attribute' on each element.

Data attributes are HTML and SVG attributes that are just used for storing data in the DOM; -- their names all
start with the string 'data-'.

Specifically, we create the - data-offset - data attribute, which we'll use to keep track of how to position each
element, and set its value to 0. -- Note that DOM attributes are always stored as strings, so the number '0' will be
converted to the string -> 0.

Next, we attach a click event handler to each element. The first thing the handler does is extract the - data-offset -
attribute from the clicked element, using - getAttribute - and store its value in the variable offset. The clicked element
is available as the target property on the event object.

Note that we use the - Number - function here to convert the string into a number. The first time this handler is called,
the variable - offset - will be set to -> 0 , as that is the initial value we store in the - data-offset - attribute.

We use the [ shiftKey ] property on the event to determine if the SHIFT key was pressed when the mouse was clicked. If it
was, we subtract 5 from -> offset. Otherwise, we add 5 to -> offset.

We then assign the updated value to the data-offset attribute using - setAttribute -

Finaly, we use the - setAttribute - method again, but this time to set the - transform - SVG attribute.