created an array of numbers to use as data
Then we create a selection of all the circle elements. The - data - method
binds the array of numbers to the selection of circles, one by one, so the
first -circle- element has the value -3- bound to it, the second -2-, and
the third -1-.

Finally, we use the - attr - method to set the radius of each circle to a
computed value based on the bound data.

As you see, if you use a function instead of a value to set an attribute,
that function will be called to compute the value for each element in the
selection.

The -d- parameter of the fucntion corresponds to the datum bound to the
current element.