updated the README.md and modified the script.js
Here we've created a longer array of numbers. We've also added a line to select
the - svg - element before selecting the - circle - elements within it.

This is necessary because D3 will need to add new circle elements, and it needs to
know which containing element to add them to.

Finally, we've added a call to the - join - method. -- This method takes the name
of the element from the selection to add or remove - match the data.

In this case, we're saying that if there aren't enough - circle - elements in the
- svg - element for all the items in -data- , then D3 should add more ( or conversely
if there are too many, D3 should remove some).

If you reload the page, you'll see this doesn't work as you probably expected. The
new circles all end up in the top-left corner of the drawing area.

That's because these new circles don't have their -cx- or -cy- attributes set, unlike
the initial that were defined in the - index.html -

To fix this, we need to set these two attributes using D3.