updated index.html and modified the script.js
The -cx- attribute is based on the index of the data. The first element
should be at 50, the second at 100, and so on.

The calculation -(i+i)*50- gives us the right values. Because the circles
are all in a line, the -cy- attribute is just a constant value.

Now when you reload the page you should see five circles in a line.

Now that we're using D3's - join - method to create new SVG elements as
needed to suit the data, there's no reason to create them in the -- HTML file.