added buttons to index.html and created a -update- function in script.js
There's no functional change here - we're just creating an -updatep function to
do the SVG updating for us, and then calling it.

Notice that we're passing -data-, the function's parameter, to the -.data- method,
rather than passing the -numbers array directly.

Next we'll add the code for handling button clicks, which will insert a random
floating-point number between 1 to 5 into the numbers array at the start or end
or drop the last element in the array.