created some event handlers and a function in script.js
First we declare a helper function for generating a random number, since
there are two places where we need to do this.

Then we declare the handlers for the three buttons. Notice that instead of using
the regular DOM API methods for adding click handlers, as we've done previously,
we're using -d3.select- to select the buttons and the -on- method to add an event
handler.

The regualr DOM API methods would work as well, but using D3 mthods is more concise
and more consistent with the other D3 code in this file.

The first handler is triggered by a click on the Append button : it pushes a random
number onto the end of the -numbers- array, then we call the -update- function to
redraw the visualization with an extra circle.

The second handler, triggered by a click on the Prepend button, causes a random
number to be unshifted onto the front of the -numbers- array.

The third is triggered by a click on the Drop button; it pops the last number from
the array.

After each of these actions we also call the -update- function.

Reload the page and try out the different buttons. You should see the elements being
added and removed as needed.