added transition to the -update- function
The -transition- method in a chain like this means that every following
attribute change will animate from its current value to the new value.

The -duration- method sets the length of the animation in milliseconds.

This means that the position and radius of each circle will take half a
second (500 ms) to animate from its current value to the new value.

New circles start off with default values of 0 for each attribute, so
they will transition in from the top-left corner of the SVG.

Unfortuantely, they way we've coded our -update- function, the animation
won't be quite as satisfying as you might want.

Reload the page and click ->Prepend<- button a few times. You should see
some odd behavior. You might have expected the existing circles to move
over to the right, making room for a new circle being added on the left.

Instead, the existing circles all appear to resize in place, while a new
circle flies in from the top-left corner and takes its place to the right
of the existing circles.

With this animation, it's actually very hard to see that an element is being
appended on the left.

Rather, the animations suggest that an element is being appended on the right,
and that all the elements are being resized.

Clicking the Append button, on the other hand, does the correct thing: a new
elements animates in and appears at the end of the row, while the existing
don't change.

The PROBLEM here is that when D3 updates an existing selection with a new array
it uses a default mode called [ join-by-index ] <-- This means that the first
item in the array is joined with the first element in the selection ( in this
case, the leftmost circle ), the second item in the array with the second element
in the selection, and so on.

If there are more items in the array than existing SVG elements, new elements are
added at the end. -- Thus, when you clik Prepend<-- and add a new number to the
start of the data array, every -circle- in the line is [ re-bound ] <-- to a new
datum.

The first circle in the line is bound to the new number that's been added to the
start of the array, so it appears to resize.

The second circle is bound to what used to be the first number in the array, so it
appears to resize as well, and so on.

## Finally, since there's now one more data item than there are SVG elements, a new
circle is created and added at the end of the line.

-> The solution to making the animation more intuitive is to help D3 understand the
-> [ identity ] <- of each element in the array of data.

Instead of assuming the every index in the array will always map to the same index
in the selection, we provide what D3 calls a [ key function ]. The key function
allows us to specify something about each datum that identifies it uniquely.

This way, each existing datum stays bound to the same SVG element even as new data
is added, regardless of the ordering of the data.