updated README.md and modified the -join- method in -update- function
This more advanced version of the -join- method takes three functions.

The first function has a single parameter called [enter], which is a selection
of temporary placeholders for each of the entering elements. To get the same
behavior as the simple [ .join(circle) ] version, we just use the -append-
method to add a circle to each -enter- placeholder.

Note that the -enter- placeholders themselves aren't elements in the DOM. They're
just a handle for D3 to give you a place to append your new entering elements,
before they get added to the DOM.

For example, if there were five new elements needing to be entered, then
[ enter.append(circle) ] would create five new circle
elements and place them inside the -svg- element.

The second function has a single, parameter called -update- which is a selection
containing all the elements that should be removed because they no longer
have a corresponding datum.

To get the same behavior as before, we call the -remove- method on the selection,
which removes each exiting element from the DOM.

When you reload the page, you should see the same behavior as before, so far,
this change doesn't have any functional impact.