updated the -.join- method
In this updated code, we've moved all the transitions into the individual
[ enter ], [ update ], and [ exit ] function.

Instead of having a single -transition- call for all the elements. The -enter-
function appends a -circle- element and then immediately sets its position
( the -cx- and -cy- attributes ), but not its radius.

Once the position is set, we use the -transition- method to animate the radius
from zero ( the default value ) to the value calculated from the datum.

The order here is important: anything that comes in the chain before the call
to -transition- will happen immediately, and the change in size ( from zero
to the desired radius ) will animate.

This will agruably look more natural than the previous version, where all three
attributes animated in from zero, leading to the circles flying from the corner.

The -update- function has to animate only the [cx] attribute of the circle
to slide it to its updated position.

All other attributes should be unchanged for existing elements.

Finally, the -exit- function animates the radius of the circle back to zero
before removing it.

If -remove- is called after a call to -transition- as it is here,
the actual element removal won't take place until after the animations
have completed.

When you reload the page, you should see the new and improved animations:

- new elements expan in at the appropriate position and removed
  elements shrink away.